pyproject.toml
[project]
name = "github-rest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
]
